requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Chemistry & Molecular Processing
rdkit>=2023.3.1

# Graph Database & Processing
neo4j>=5.0.0
graphdatascience>=1.5.0

# Deep Learning (PyTorch)
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.16
torch-cluster>=1.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
scipy>=1.11.0

# Jupyter Notebook
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.0.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
